kernel: Ignore error when setting function filter
The skip_kernel_functions() adds a list of functions which seems not so
useful for uftrace to the filter list.  But some of them might not exist
in the kernel due to version or configuration differences.

This is especially annoying if the system doesn't provide vDSO for
clock_gettime() function which is called from uftrace/libmcount.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>